content-service/requirements.txt
fastapi
uvicorn

pika

python-dotenv
supabase